+++ lottery-core/src/main/java/de/alpharogroup/lottery/drawing/DrawnLotteryNumbersExtensions.java
@@ -23,25 +23,25 @@ package de.alpharogroup.lottery.drawing;
 import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.logging.Level;
+import java.util.stream.Collectors;
 
 import de.alpharogroup.collections.list.ListFactory;
-import de.alpharogroup.collections.map.MapFactory;
+import de.alpharogroup.collections.map.MapExtensions;
 import de.alpharogroup.collections.set.SetFactory;
+import de.alpharogroup.comparators.ComparatorFactory;
 import de.alpharogroup.random.DefaultSecureRandom;
 import de.alpharogroup.random.number.RandomPrimitivesExtensions;
-import lombok.SneakyThrows;
 import lombok.experimental.UtilityClass;
-import lombok.extern.java.Log;
 
 /**
- * The class {@link DrawnLotteryNumbersExtensions}.
+ * The class {@link DrawnLotteryNumbersExtensions} provides utility methods to draw lottery numbers
+ * with different algorithms
  */
 @UtilityClass
-@Log
 public final class DrawnLotteryNumbersExtensions
 {
 
@@ -54,7 +54,6 @@ public final class DrawnLotteryNumbersExtensions
 	 *            the volume of the numbers starts from 1 till volume
 	 * @return the drawn super number
 	 */
-	@SneakyThrows
 	public static int drawSuperNumber(Set<Integer> alreadyDrawnNumbers, int volume)
 	{
 		int superNumber = -1;
@@ -81,7 +80,6 @@ public final class DrawnLotteryNumbersExtensions
 	 *            the max volume
 	 * @return the drawn super number
 	 */
-	@SneakyThrows
 	public static int drawSuperNumber(Set<Integer> alreadyDrawnNumbers, int minVolume,
 		int maxVolume)
 	{
@@ -118,7 +116,6 @@ public final class DrawnLotteryNumbersExtensions
 	 *            the volume of the numbers starts from 1 till volume
 	 * @return the sets the
 	 */
-	@SneakyThrows
 	public static Set<Integer> draw(int maxNumbers, int volume)
 	{
 		Set<Integer> numbers = SetFactory.newTreeSet();
@@ -186,41 +183,122 @@ public final class DrawnLotteryNumbersExtensions
 	public static Set<Integer> drawFromMultiMap(int maxNumbers, int minVolume, int maxVolume,
 		int drawCount)
 	{
-		Map<Integer, Integer> numberCount = MapFactory.newHashMap();
-		for (int i = minVolume; i <= maxVolume; i++)
-		{
-			numberCount.put(i, 0);
-		}
-		for (int i = 0; i < drawCount; i++)
-		{
-			Set<Integer> lotteryNumbers = DrawnLotteryNumbersExtensions.draw(maxNumbers, minVolume,
-				maxVolume);
-
-			lotteryNumbers.stream().forEach(key -> numberCount.merge(key, 1, Integer::sum));
-		}
-		List<Map.Entry<Integer, Integer>> sortByValue = sortByValue(numberCount);
-		List<Integer> newLotteryNumbers = ListFactory.newArrayList();
-		int count = 1;
-		for (Map.Entry<Integer, Integer> entry : sortByValue)
-		{
-			if (maxNumbers < count)
-			{
-				break;
-			}
-			newLotteryNumbers.add(entry.getKey());
-			count++;
-		}
-		return SetFactory.newTreeSet(newLotteryNumbers);
+		return drawFromMultiMap(maxNumbers, minVolume, maxVolume, drawCount, true);
 	}
 
-
-	public static <K, V extends Comparable<? super V>> List<Map.Entry<K, V>> sortByValue(
-		Map<K, V> map)
+	/**
+	 * Draw of paranoid lottery numbers from given drawCount and take the numbers that are drawn the
+	 * most times and return a new set.
+	 *
+	 * @param maxNumbers
+	 *            the maximum of numbers to draw
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @param drawCount
+	 *            the draw count defines how many times to draw numbers
+	 * @return the sets of the drawn numbers
+	 */
+	public static Set<Integer> drawParanoidFromMultiMap(int maxNumbers, int minVolume,
+		int maxVolume, int drawCount)
 	{
-		List<Map.Entry<K, V>> list = ListFactory.newArrayList(map.entrySet());
-		list.sort(Map.Entry.comparingByValue());
-		Collections.reverse(list);
-		return list;
+		return drawFromMultiMap(maxNumbers, minVolume, maxVolume, drawCount,
+			RandomPrimitivesExtensions.randomBoolean(), true);
+	}
+
+	/**
+	 * Draw of lottery numbers from given drawCount and take the numbers that are drawn the most
+	 * times and return a new set.
+	 *
+	 * @param maxNumbers
+	 *            the maximum of numbers to draw
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @param drawCount
+	 *            the draw count defines how many times to draw numbers
+	 * @param mostDrawn
+	 *            the flag that indicates if the most drawn numbers should be taken if true,
+	 *            otherwise the reverse order will be taken
+	 * @return the sets of the drawn numbers
+	 */
+	public static Set<Integer> drawFromMultiMap(int maxNumbers, int minVolume, int maxVolume,
+		int drawCount, boolean mostDrawn)
+	{
+		return drawFromMultiMap(maxNumbers, minVolume, maxVolume, drawCount, mostDrawn, false);
+	}
+
+	/**
+	 * Draw of lottery numbers from given drawCount and take the numbers that are drawn the most
+	 * times and return a new set.
+	 *
+	 * @param maxNumbers
+	 *            the maximum of numbers to draw
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @param drawCount
+	 *            the draw count defines how many times to draw numbers
+	 * @param mostDrawn
+	 *            the flag that indicates if the most drawn numbers should be taken if true,
+	 *            otherwise the reverse order will be taken
+	 * @param paranoid
+	 *            the flag paranoid indicates to create a custom comparator from the counter map and
+	 *            define a random defined order to draw if true, otherwise the flag mostDrawn will
+	 *            define the order to draw
+	 * @return the sets of the drawn numbers
+	 */
+	public static Set<Integer> drawFromMultiMap(int maxNumbers, int minVolume, int maxVolume,
+		int drawCount, boolean mostDrawn, boolean paranoid)
+	{
+		Map<Integer, Integer> numberCounterMap = DrawnLotteryNumbersFactory
+			.newNumberCounterMap(minVolume, maxVolume);
+		for (int i = 0; i < drawCount; i++)
+		{
+			DrawnLotteryNumbersExtensions.draw(maxNumbers, minVolume, maxVolume)
+				.forEach(key -> numberCounterMap.merge(key, 1, Integer::sum));
+		}
+		Comparator<Integer> mostDrawnComparator;
+		if (paranoid)
+		{
+			List<Integer> numberCounterValues = ListFactory
+				.newArrayList(SetFactory.newTreeSet(numberCounterMap.values()));
+			Collections.shuffle(numberCounterValues, DefaultSecureRandom.get());
+			mostDrawnComparator = ComparatorFactory.newComparator(numberCounterValues);
+		}
+		else
+		{
+			mostDrawnComparator = mostDrawn ? Comparator.reverseOrder() : Comparator.naturalOrder();
+		}
+		return resolveLotteryNumbers(maxNumbers, mostDrawnComparator, numberCounterMap);
+	}
+
+	/**
+	 * Resolves the lottery numbers from the given number counter map in the order from the given
+	 * comparator limited to maxNumbers
+	 *
+	 * @param maxNumbers
+	 *            the max numbers
+	 * @param mostDrawn
+	 *            the comparator that defines in which order to take the drawn numbers. For instance
+	 *            if you want to have the reverse order you can simply give the
+	 *            <code>Comparator.reverseOrder()</code> or you can define your custom order
+	 * @param numberCounterMap
+	 *            the number counter map
+	 * @return the sets of the lottery numbers
+	 */
+	public static Set<Integer> resolveLotteryNumbers(int maxNumbers, Comparator<Integer> mostDrawn,
+		Map<Integer, Integer> numberCounterMap)
+	{
+		List<Map.Entry<Integer, Integer>> sortByValue = MapExtensions
+			.sortByValueAsList(numberCounterMap, mostDrawn);
+
+		List<Integer> newLotteryNumbers = sortByValue.stream().map(Map.Entry::getKey)
+			.limit(maxNumbers).collect(Collectors.toList());
+		return SetFactory.newTreeSet(newLotteryNumbers);
 	}
 
 	/**
@@ -234,7 +312,6 @@ public final class DrawnLotteryNumbersExtensions
 	 *            the volume of the numbers starts from 1 till volume
 	 * @return the sets the
 	 */
-	@SneakyThrows
 	public static Set<Integer> drawDefaultAlgorithm(int maxNumbers, int minVolume, int maxVolume)
 	{
 		Set<Integer> numbers = SetFactory.newTreeSet();
@@ -250,10 +327,6 @@ public final class DrawnLotteryNumbersExtensions
 			final int index = RandomPrimitivesExtensions.randomIntBetween(0, rangeList.size(), true,
 				false);
 			Integer drawnNumber = rangeList.get(index);
-			if (rangeList.remove(drawnNumber))
-			{
-				log.log(Level.FINE, "removed drawn number:" + drawnNumber);
-			}
 			if (!numbers.contains(drawnNumber))
 			{
 				numbers.add(drawnNumber);