+++ lottery-core/src/main/java/de/alpharogroup/lottery/drawing/DrawnLotteryNumbersFactory.java
@@ -21,31 +21,27 @@
 package de.alpharogroup.lottery.drawing;
 
 import java.security.SecureRandom;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
 
-import de.alpharogroup.collections.list.ListFactory;
+import de.alpharogroup.collections.map.MapFactory;
 import de.alpharogroup.collections.set.SetFactory;
 import de.alpharogroup.lottery.drawings.DrawnLotteryNumbers;
 import de.alpharogroup.lottery.enums.LotteryAlgorithm;
 import de.alpharogroup.random.SecureRandomFactory;
 import de.alpharogroup.random.number.RandomPrimitivesExtensions;
 import lombok.NonNull;
-import lombok.SneakyThrows;
 import lombok.experimental.UtilityClass;
 
 /**
- * A factory for creating {@link DrawnLotteryNumbers} objects with generated lottery numbers.
+ * A factory for creating {@link DrawnLotteryNumbers} objects with generated lottery numbers
  */
 @UtilityClass
-public class DrawnLotteryNumbersFactory
+public final class DrawnLotteryNumbersFactory
 {
 
 	/**
-	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers.
+	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers
 	 *
 	 * @param max
 	 *            the max number to draw
@@ -83,7 +79,7 @@ public class DrawnLotteryNumbersFactory
 	}
 
 	/**
-	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers.
+	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers
 	 *
 	 * @param max
 	 *            the max number to draw
@@ -93,7 +89,6 @@ public class DrawnLotteryNumbersFactory
 	 *            the max volume
 	 * @return the new {@link DrawnLotteryNumbers}
 	 */
-	@SneakyThrows
 	public static DrawnLotteryNumbers newRandomDrawnLotteryNumbers(int max, int minVolume,
 		int maxVolume)
 	{
@@ -120,17 +115,39 @@ public class DrawnLotteryNumbersFactory
 	 *            the algorithm to use
 	 * @return the new {@link DrawnLotteryNumbers}
 	 */
-	@SneakyThrows
 	public static DrawnLotteryNumbers newRandomDrawnLotteryNumbers(int max, int minVolume,
 		int maxVolume, @NonNull LotteryAlgorithm algorithm)
+	{
+		return newRandomDrawnLotteryNumbers(max, minVolume, maxVolume, 200, algorithm);
+	}
+
+	/**
+	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers
+	 * with the given algorithm
+	 *
+	 * @param max
+	 *            the max number to draw
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @param drawCount
+	 *            the draw count defines how many times to draw numbers. Note: only with map
+	 *            algorithm
+	 * @param algorithm
+	 *            the algorithm to use
+	 * @return the new {@link DrawnLotteryNumbers}
+	 */
+	public static DrawnLotteryNumbers newRandomDrawnLotteryNumbers(int max, int minVolume,
+		int maxVolume, int drawCount, @NonNull LotteryAlgorithm algorithm)
 	{
 		switch (algorithm)
 		{
 			case MAP :
 				DrawnLotteryNumbers drawnLotteryNumbers = newRandomDrawnLotteryNumbers(max,
 					minVolume, maxVolume);
-				drawnLotteryNumbers.setLotteryNumbers(
-					DrawnLotteryNumbersExtensions.drawFromMultiMap(max, minVolume, maxVolume, 200));
+				drawnLotteryNumbers.setLotteryNumbers(DrawnLotteryNumbersExtensions
+					.drawFromMultiMap(max, minVolume, maxVolume, drawCount));
 				return drawnLotteryNumbers;
 			case SET :
 				return newRandomDrawnLotteryNumbers(max, maxVolume);
@@ -140,16 +157,8 @@ public class DrawnLotteryNumbersFactory
 		}
 	}
 
-	public static <K, V extends Comparable<? super V>> List<Entry<K, V>> sortByValue(Map<K, V> map)
-	{
-		List<Entry<K, V>> list = ListFactory.newArrayList(map.entrySet());
-		list.sort(Entry.comparingByValue());
-		Collections.reverse(list);
-		return list;
-	}
-
 	/**
-	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers.
+	 * Factory method for create a new {@link DrawnLotteryNumbers} object with all drawn numbers
 	 *
 	 * @param max
 	 *            the max number to draw
@@ -164,12 +173,8 @@ public class DrawnLotteryNumbersFactory
 		int id = RandomPrimitivesExtensions.randomInt(Integer.MAX_VALUE);
 		int superNumber = DrawnLotteryNumbersExtensions.drawSuperNumber(lotteryNumbers, volume);
 		int superSixNumber = RandomPrimitivesExtensions.randomIntBetween(1, 10);
-
-		final DrawnLotteryNumbers drawnLotteryNumbers = DrawnLotteryNumbers.builder().id(id)
-			.lotteryNumbers(lotteryNumbers).superNumber(superNumber).superSixNumber(superSixNumber)
-			.build();
-
-		return drawnLotteryNumbers;
+		return DrawnLotteryNumbers.builder().id(id).lotteryNumbers(lotteryNumbers)
+			.superNumber(superNumber).superSixNumber(superSixNumber).build();
 	}
 
 	/**
@@ -180,6 +185,25 @@ public class DrawnLotteryNumbersFactory
 	public static DrawnLotteryNumbers newRandomDrawnLotteryNumbers()
 	{
 		return newRandomDrawnLotteryNumbers(7, 49);
+	}
+
+	/**
+	 * Factory method for create a map for count drawn numbers
+	 *
+	 * @param minVolume
+	 *            the min volume
+	 * @param maxVolume
+	 *            the max volume
+	 * @return the new map with the initial values
+	 */
+	public static Map<Integer, Integer> newNumberCounterMap(int minVolume, int maxVolume)
+	{
+		Map<Integer, Integer> numberCount = MapFactory.newHashMap();
+		for (int i = minVolume; i <= maxVolume; i++)
+		{
+			numberCount.put(i, 0);
+		}
+		return numberCount;
 	}
 
 }