+++ lottery-core/src/test/java/de/alpharogroup/lottery/drawing/DrawnLotteryNumbersExtensionsTest.java
@@ -20,6 +20,8 @@
  */
 package de.alpharogroup.lottery.drawing;
 
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
 import java.lang.reflect.InvocationTargetException;
@@ -44,6 +46,8 @@ public class DrawnLotteryNumbersExtensionsTest
 	@Test
 	public void testDrawDefaultAlgorithm()
 	{
+		int actual;
+		int expected;
 
 		Set<Integer> lotteryNumbers;
 		int max;
@@ -57,7 +61,38 @@ public class DrawnLotteryNumbersExtensionsTest
 		maxVolume = 39;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.drawDefaultAlgorithm(max, minVolume,
 			maxVolume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
+		for (Integer lotteryNumber : lotteryNumbers)
+		{
+			assertTrue(MathExtensions.isBetween(minVolume, maxVolume, lotteryNumber, true, true));
+		}
+	}
+
+	/**
+	 * Test method for {@link DrawnLotteryNumbersExtensions#drawFromMultiMap(int, int, int, int)}
+	 */
+	@Test
+	public void testDrawFromMultiMap()
+	{
+		int actual;
+		int expected;
+		Set<Integer> lotteryNumbers;
+		int maxNumbers;
+		int minVolume;
+		int maxVolume;
+		int drawCount;
+
+		maxNumbers = 7;
+		minVolume = 1;
+		maxVolume = 70;
+		drawCount = 10000;
+		lotteryNumbers = DrawnLotteryNumbersExtensions.drawFromMultiMap(maxNumbers, minVolume,
+			maxVolume, drawCount);
+		actual = lotteryNumbers.size();
+		expected = maxNumbers;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(minVolume, maxVolume, lotteryNumber, true, true));
@@ -81,6 +116,8 @@ public class DrawnLotteryNumbersExtensionsTest
 	@Test
 	public void testDrawIntInt()
 	{
+		int actual;
+		int expected;
 		Set<Integer> lotteryNumbers;
 		int volume;
 		int max;
@@ -90,7 +127,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 7 of 39
 		volume = 39;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -99,7 +138,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 7 of 36
 		volume = 36;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -108,7 +149,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 7 of 35
 		volume = 35;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -119,7 +162,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 59
 		volume = 59;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -128,7 +173,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 49
 		volume = 49;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -137,7 +184,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 48
 		volume = 48;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -146,7 +195,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 45
 		volume = 45;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -155,7 +206,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 42
 		volume = 42;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -164,7 +217,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 6 of 30
 		volume = 30;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -175,7 +230,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 90
 		volume = 90;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -184,7 +241,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 75
 		volume = 75;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -193,7 +252,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 69
 		volume = 69;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -202,7 +263,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 49
 		volume = 49;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -211,7 +274,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 42
 		volume = 42;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -220,7 +285,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 40
 		volume = 40;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -229,7 +296,9 @@ public class DrawnLotteryNumbersExtensionsTest
 		// simulate a draw of 5 of 35
 		volume = 35;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, volume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(1, volume, lotteryNumber, true, true));
@@ -242,6 +311,8 @@ public class DrawnLotteryNumbersExtensionsTest
 	@Test
 	public void testDrawIntIntInt()
 	{
+		int actual;
+		int expected;
 		Set<Integer> lotteryNumbers;
 		int max;
 		int minVolume;
@@ -253,7 +324,39 @@ public class DrawnLotteryNumbersExtensionsTest
 		minVolume = 5;
 		maxVolume = 39;
 		lotteryNumbers = DrawnLotteryNumbersExtensions.draw(max, minVolume, maxVolume);
-		assertTrue(lotteryNumbers.size() == max);
+		actual = lotteryNumbers.size();
+		expected = max;
+		assertEquals(actual, expected);
+		for (Integer lotteryNumber : lotteryNumbers)
+		{
+			assertTrue(MathExtensions.isBetween(minVolume, maxVolume, lotteryNumber, true, true));
+		}
+	}
+
+	/**
+	 * Test method for
+	 * {@link DrawnLotteryNumbersExtensions#drawParanoidFromMultiMap(int, int, int, int)}
+	 */
+	@Test
+	public void testDrawParanoidFromMultiMap()
+	{
+		int actual;
+		int expected;
+		Set<Integer> lotteryNumbers;
+		int maxNumbers;
+		int minVolume;
+		int maxVolume;
+		int drawCount;
+
+		maxNumbers = 7;
+		minVolume = 1;
+		maxVolume = 70;
+		drawCount = 1000;
+		lotteryNumbers = DrawnLotteryNumbersExtensions.drawParanoidFromMultiMap(maxNumbers,
+			minVolume, maxVolume, drawCount);
+		actual = lotteryNumbers.size();
+		expected = maxNumbers;
+		assertEquals(actual, expected);
 		for (Integer lotteryNumber : lotteryNumbers)
 		{
 			assertTrue(MathExtensions.isBetween(minVolume, maxVolume, lotteryNumber, true, true));
@@ -268,7 +371,7 @@ public class DrawnLotteryNumbersExtensionsTest
 	{
 		Set<Integer> alreadyDrawnNumbers = SetFactory.newTreeSet(6, 12, 15, 18, 25, 30);
 		int superNumber = DrawnLotteryNumbersExtensions.drawSuperNumber(alreadyDrawnNumbers, 49);
-		assertTrue(!alreadyDrawnNumbers.contains(superNumber));
+		assertFalse(alreadyDrawnNumbers.contains(superNumber));
 		assertTrue(MathExtensions.isBetween(1, 49, superNumber, true, true));
 	}